% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagDeviations.R
\docType{methods}
\name{bagDeviations}
\alias{bagDeviations}
\alias{bagDeviations,chromVARDeviations,numeric,character-method}
\title{bagDeviations}
\usage{
bagDeviations(object, cor, organism)

\S4method{bagDeviations}{chromVARDeviations,numeric,character}(object,
  cor = 0.7, organism = "human")
}
\arguments{
\item{object}{chromVARDeviations object from
chromVAR after}

\item{cor}{= 0.7 minimum correlation between PWMS}

\item{organism}{= "human"; only other option is "mouse"}
}
\description{
Given a deviations object from the cisBP
curtailed chromVARmotifs list, "bag" them
according to correlation where the most variable
goes first. Make sure to specify organism and
correlation. The lower the the correlation, 
the less bagging that occurs. Row data then
contains a comma-separated list of motifs that are
bagged.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = chromVARDeviations,cor = numeric,organism = character}: 
}}

\examples{

rdsA<-paste0(system.file('rds',package='chromVARxx'),'/dev_humanSimple.rds')
object <- readRDS(rdsA)
bagged <- bagDeviations(object, cor = 0.3, organism = "human")

}
\author{
Caleb Lareau
}
