% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buenVAR.R
\docType{package}
\name{buenVAR}
\alias{buenVAR}
\alias{buenVAR-package}
\title{buenVAR: Buenrostro Lab chromVAR extensions}
\description{
The buenVAR package contains functions to augment
chromVAR functionality designed in the Buenrostro
lab
}
