% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weighted_deviations.R
\docType{methods}
\name{computeWeightedDeviations}
\alias{computeWeightedDeviations}
\alias{computeWeightedDeviations,SummarizedExperiment,SummarizedExperiment-method}
\alias{computeWeightedDeviations,SummarizedExperiment,MatrixOrmatrix-method}
\alias{computeWeightedDeviations,SummarizedExperiment,list-method}
\alias{computeWeightedDeviations,SummarizedExperiment,missingOrNULL-method}
\alias{computeWeightedDeviations,MatrixOrmatrix,SummarizedExperiment-method}
\alias{computeWeightedDeviations,MatrixOrmatrix,MatrixOrmatrix-method}
\alias{computeWeightedDeviations,MatrixOrmatrix,list-method}
\alias{computeWeightedDeviations,MatrixOrmatrix,missingOrNULL-method}
\title{computeWeightedDeviations}
\usage{
computeWeightedDeviations(object, annotations, ...)


  \S4method{computeWeightedDeviations}{SummarizedExperiment,SummarizedExperiment}(object,
  annotations, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))


  \S4method{computeWeightedDeviations}{SummarizedExperiment,MatrixOrmatrix}(object,
  annotations, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))

\S4method{computeWeightedDeviations}{SummarizedExperiment,list}(object,
  annotations, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))


  \S4method{computeWeightedDeviations}{SummarizedExperiment,missingOrNULL}(object,
  annotations, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))


  \S4method{computeWeightedDeviations}{MatrixOrmatrix,SummarizedExperiment}(object,
  annotations, background_peaks, expectation = computeExpectations(object))

\S4method{computeWeightedDeviations}{MatrixOrmatrix,MatrixOrmatrix}(object,
  annotations, background_peaks, expectation = computeExpectations(object))

\S4method{computeWeightedDeviations}{MatrixOrmatrix,list}(object, annotations,
  background_peaks, expectation = computeExpectations(object))

\S4method{computeWeightedDeviations}{MatrixOrmatrix,missingOrNULL}(object,
  annotations, background_peaks, expectation = computeExpectations(object))
}
\arguments{
\item{object}{chromVARCounts object}

\item{annotations}{chromVARAnnotations object}

\item{...}{additional arguments}

\item{background_peaks}{(optional) background peaks matrix computed using
\code{\link{getBackgroundPeaks}}, computed internally with default
paramaters if not provided}

\item{expectation}{(optional) expectations computed using
\code{\link{computeExpectations}}, computed automatically if not provided}
}
\value{
\code{\link{chromVARDeviations-class}}, which inherits from
SummarizedExperiment, and has two assays: deviations and deviation scores.
}
\description{
Computes weighted deviations in chromatin accessibility
across sets of annotations where the number of the annotation
meaningful.
}
\details{
multiprocessing using \code{\link[BiocParallel]{bplapply}}
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SummarizedExperiment,annotations = SummarizedExperiment}: / annotations are SummarizedExperiment

\item \code{object = SummarizedExperiment,annotations = MatrixOrmatrix}: object is SummarizedExperiment,
annotations are Matrix

\item \code{object = SummarizedExperiment,annotations = list}: object is SummarizedExperiment,
annotations are list

\item \code{object = SummarizedExperiment,annotations = missingOrNULL}: object is SummarizedExperiment,
annotations are missing

\item \code{object = MatrixOrmatrix,annotations = SummarizedExperiment}: object and annotations are SummarizedExperiment

\item \code{object = MatrixOrmatrix,annotations = MatrixOrmatrix}: object is SummarizedExperiment,
annotations are Matrix

\item \code{object = MatrixOrmatrix,annotations = list}: object is SummarizedExperiment,
annotations are list

\item \code{object = MatrixOrmatrix,annotations = missingOrNULL}: object is SummarizedExperiment,
annotations are missing
}}

\examples{
# Register BiocParallel
BiocParallel::register(BiocParallel::SerialParam())
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
# load annotation matrix; result from matchMotifs
data(mini_ix, package = "chromVAR")

# computing deviations
dev <- computeWeightedDeviations(object = mini_counts,
                         annotations = mini_ix)
}
\seealso{
\code{\link{computeVariability}}, \code{\link{plotVariability}}
}
\author{
Caleb Lareau, Alicia Schep
}
