% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gchromVAR.R
\docType{package}
\name{gchromVAR}
\alias{gchromVAR}
\alias{gchromVAR-package}
\title{gchromVAR: High-resolution cell type enrichments for 
epigentic and fine-mapped GWAS variants}
\description{
The gchromVAR package implements a bias-corrected weighted
deviation function and other supporting functions
to infer a high-resolution enrichment of finemapped
variants in their most probably causal tissue.
}
