% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{importBedScore}
\alias{importBedScore}
\alias{importBedScore,GRanges,character-method}
\title{importBedScore}
\usage{
importBedScore(ranges, files, colidx = 5, FUN = sum, default.val = 0)

\S4method{importBedScore}{GRanges,character}(ranges, files, colidx = 5,
  FUN = sum, default.val = 0)
}
\arguments{
\item{ranges}{A GRanges object corresponding to the
peaks used to aggre}

\item{files}{}
}
\description{
Given a set of peaks (GRanges object) and a vector
of one or more file paths of .bed files with some
score,
}
\section{Methods (by class)}{
\itemize{
\item \code{ranges = GRanges,files = character}: 
}}

\examples{

files <- list.files(system.file('extdata',package='chromVARxx'), full.names = TRUE)
data(mini_counts, package = "chromVAR")
w_se <- importBedScore(rowRanges(mini_counts), files)

}
\author{
Caleb Lareau
}
