% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWeightedDeviations.R
\docType{methods}
\name{computeWeightedDeviations}
\alias{computeWeightedDeviations}
\alias{computeWeightedDeviations,SummarizedExperiment,SummarizedExperiment-method}
\alias{computeWeightedDeviations,SummarizedExperiment,Matrix-method}
\alias{computeWeightedDeviations,SummarizedExperiment,matrix-method}
\title{computeWeightedDeviations}
\usage{
computeWeightedDeviations(object, weights, ...)


  \S4method{computeWeightedDeviations}{SummarizedExperiment,SummarizedExperiment}(object,
  weights, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))

\S4method{computeWeightedDeviations}{SummarizedExperiment,Matrix}(object,
  weights, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))

\S4method{computeWeightedDeviations}{SummarizedExperiment,matrix}(object,
  weights, background_peaks = getBackgroundPeaks(object),
  expectation = computeExpectations(object))
}
\arguments{
\item{object}{chromVARCounts object}

\item{weights}{chromVARweights object}

\item{...}{additional arguments}

\item{background_peaks}{(optional) background peaks matrix computed using
\code{\link{getBackgroundPeaks}}, computed internally with default
paramaters if not provided}

\item{expectation}{(optional) expectations computed using
\code{\link{computeExpectations}}, computed automatically if not provided}
}
\value{
\code{\link{chromVARDeviations-class}}, which inherits from
SummarizedExperiment, and has two assays: deviations and deviation scores.
}
\description{
Computes weighted deviations in chromatin accessibility
across sets of weights where the number of the annotation
meaningful.
}
\details{
multiprocessing using \code{\link[BiocParallel]{bplapply}}
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SummarizedExperiment,weights = SummarizedExperiment}: / weights are SummarizedExperiment

\item \code{object = SummarizedExperiment,weights = Matrix}: object is SummarizedExperiment,
weights are in a Matrix

\item \code{object = SummarizedExperiment,weights = matrix}: object is SummarizedExperiment,
weights are in a numeric
}}

\examples{
# Register BiocParallel
BiocParallel::register(BiocParallel::SerialParam())
# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
# Load mini weighted counts
rdsA<-paste0(system.file('rds',package='chromVARxx'),'/mini_w.rds')
w_se <- readRDS(rdsA)

# Build weights from .bed file
files <- list.files(system.file('extdata',package='chromVARxx'), full.names = TRUE)
data(mini_counts, package = "chromVAR")
uk_se <- importBedScore(SummarizedExperiment::rowRanges(mini_counts), files)

wdev <- computeWeightedDeviations(mini_counts, w_se)
ukdev <- computeWeightedDeviations(mini_counts, uk_se)

}
\author{
Caleb Lareau
}
