% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{methods}
\name{importBedScore}
\alias{importBedScore}
\alias{importBedScore,GRanges,character-method}
\title{importBedScore}
\usage{
importBedScore(ranges, files, colidx = 5, FUN = sum, default.val = 0)

\S4method{importBedScore}{GRanges,character}(ranges, files, colidx = 5,
  FUN = sum, default.val = 0)
}
\arguments{
\item{ranges}{A GRanges object corresponding to the
peaks used to aggregrate a score over}

\item{files}{A character vector of files that will
be imported and digested for analysis}

\item{colidx}{The column index of the score. This
assumes that the first three columns specific genomic 
coordinates. Default is 5 for the fifth column in the .bed files}

\item{FUN}{Function to summarize multiple hits in the
.bed file over the peak.}

\item{default.val}{Default value to populate the matrices
ahead of time. By default, 0.}
}
\description{
Given a set of peaks (GRanges object) and a vector
of one or more file paths of .bed files with some
score, import these and summarize using a FUN
over them. Also can specify which column the score 
should be read from in the file and the default
value per peak if nothing is determined.
}
\section{Methods (by class)}{
\itemize{
\item \code{ranges = GRanges,files = character}: 
}}

\examples{

files <- list.files(system.file('extdata',package='chromVARxx'), full.names = TRUE)
data(mini_counts, package = "chromVAR")
w_se <- importBedScore(SummarizedExperiment::rowRanges(mini_counts), files)

}
\author{
Caleb Lareau
}
